Refactoring to improve naming, fixing issue with Basic Auth
