+++ src/main/java/com/musictrainer/api/MusictrainerApplication.java
@@ -1,0 +1,13 @@
+package com.musictrainer.api;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class MusictrainerApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(MusictrainerApplication.class, args);
+	}
+
+}
