+++ src/main/java/com/musictrainer/api/messaging/bean/TestWebServiceBean.java
@@ -1,0 +1,27 @@
+package com.musictrainer.api.messaging.bean;
+
+public class TestWebServiceBean {
+    private String message;
+
+    public TestWebServiceBean (String message){
+        System.out.println("message in bean Constructor: " + message);
+        this.message = message;
+    }
+
+    public void setMessage(String message) {
+        this.message = message;
+    }
+    @Override
+    public String toString() {
+//        return "TestWebServiceBean{" +
+//                "message='" + message + '\'' +
+//                '}';
+        return String.format("HellowWorldBean [message=%s]", message);
+    }
+    public String getMessage() {
+        return message;
+    }
+
+
+
+}
