+++ src/main/java/com/musictrainer/api/security/basicauth/BasicAuthenticationController.java
@@ -1,0 +1,19 @@
+package com.musictrainer.api.security.basicauth;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@Slf4j
+@RestController
+@CrossOrigin(origins = "http://localhost:4500")
+public class BasicAuthenticationController {
+
+    @GetMapping("/basicauth")
+    public AuthenticationBean helloWorldBean() {
+        log.info("in Basic Auth Controller");
+        return new AuthenticationBean("You are authenticated");
+    }
+}
+
