--- src/main/java/com/musictrainer/background/musictrainer/MusictrainerApplication.java
@@ -1,13 +1,0 @@
-package com.musictrainer.background.musictrainer;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class MusictrainerApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(MusictrainerApplication.class, args);
-	}
-
-}
