--- src/main/java/com/musictrainer/background/musictrainer/jwt/JwtInMemoryUserDetailsService.java
@@ -1,44 +1,0 @@
-package com.musictrainer.background.musictrainer.jwt;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-
-
-@Service
-public class JwtInMemoryUserDetailsService implements UserDetailsService {
-
-  private final Logger logger = LoggerFactory.getLogger(this.getClass());
-
-  static List<JwtUserDetails> inMemoryUserList = new ArrayList<>();
-
-  static {
-    inMemoryUserList.add(new JwtUserDetails(1L, "in28minutes",
-        "$2a$10$3zHzb.Npv1hfZbLEU5qsdOju/tk2je6W6PnNnY.c1ujWPcZh4PL6e", "ROLE_USER_2"));
-  }
-
-  @Override
-  public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-
-    logger.warn("JwtInMemoryUserDetailsService.loadUserByUsername : " + username);
-    Optional<JwtUserDetails> findFirst = inMemoryUserList.stream()
-        .filter(user -> user.getUsername().equals(username)).findFirst();
-
-    if (!findFirst.isPresent()) {
-      throw new UsernameNotFoundException(String.format("USER_NOT_FOUND '%s'.", username));
-    }
-
-    return findFirst.get();
-  }
-
-}
-
-
