--- src/main/java/com/musictrainer/background/musictrainer/jwt/JwtUnAuthorizedResponseAuthenticationEntryPoint.java
@@ -1,33 +1,0 @@
-package com.musictrainer.background.musictrainer.jwt;
-
-//import com.musictrainer.background.musictrainer.jwt.resource.AuthenticationException;
-
-import java.io.IOException;
-import java.io.Serializable;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.AuthenticationEntryPoint;
-import org.springframework.stereotype.Component;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-@Component
-public class JwtUnAuthorizedResponseAuthenticationEntryPoint implements AuthenticationEntryPoint, Serializable {
-
-  private final Logger logger = LoggerFactory.getLogger(this.getClass());
-
-  private static final long serialVersionUID = -8970718410437077606L;
-
-  @Override
-  public void commence(HttpServletRequest request, HttpServletResponse response,
-                       AuthenticationException authException) throws IOException {
-
-    logger.warn("JwtUnAuthorizedResponseAuthenticationEntryPoint.Commence");
-    response.sendError(HttpServletResponse.SC_UNAUTHORIZED,
-        "You would need to provide the Jwt Token to Access This resource");
-  }
-}
