--- src/main/java/com/musictrainer/background/musictrainer/jwt/JwtUserDetails.java
@@ -1,86 +1,0 @@
-package com.musictrainer.background.musictrainer.jwt;
-
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
-
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-
-public class JwtUserDetails implements UserDetails {
-
-  private final Logger logger = LoggerFactory.getLogger(this.getClass());
-
-  private static final long serialVersionUID = 5155720064139820502L;
-
-  private final Long id;
-  private final String username;
-  private final String password;
-  private final Collection<? extends GrantedAuthority> authorities;
-
-  public JwtUserDetails(Long id, String username, String password, String role) {
-
-    logger.warn("JwtUserDetails");
-
-    this.id = id;
-    this.username = username;
-    this.password = password;
-
-    List<SimpleGrantedAuthority> authorities = new ArrayList<SimpleGrantedAuthority>();
-    authorities.add(new SimpleGrantedAuthority(role));
-
-    this.authorities = authorities;
-  }
-
-  @JsonIgnore
-  public Long getId() {
-    return id;
-  }
-
-  @Override
-  public String getUsername() {
-    return username;
-  }
-
-  @JsonIgnore
-  @Override
-  public boolean isAccountNonExpired() {
-    return true;
-  }
-
-  @JsonIgnore
-  @Override
-  public boolean isAccountNonLocked() {
-    return true;
-  }
-
-  @JsonIgnore
-  @Override
-  public boolean isCredentialsNonExpired() {
-    return true;
-  }
-
-  @JsonIgnore
-  @Override
-  public String getPassword() {
-    return password;
-  }
-
-  @Override
-  public Collection<? extends GrantedAuthority> getAuthorities() {
-    return authorities;
-  }
-
-  @Override
-  public boolean isEnabled() {
-    return true;
-  }
-
-}
