--- src/main/java/com/musictrainer/background/musictrainer/jwt/resource/AuthenticationException.java
@@ -1,7 +1,0 @@
-package com.musictrainer.background.musictrainer.jwt.resource;
-
-public class AuthenticationException extends RuntimeException {
-    public AuthenticationException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
