--- src/main/java/com/musictrainer/background/musictrainer/jwt/resource/JwtTokenRequest.java
@@ -1,37 +1,0 @@
-package com.musictrainer.background.musictrainer.jwt.resource;
-
-import java.io.Serializable;
-
-public class  JwtTokenRequest implements Serializable {
-
-  private static final long serialVersionUID = -5616176897013108345L;
-
-  private String username;
-    private String password;
-
-    public JwtTokenRequest() {
-        super();
-    }
-
-    public JwtTokenRequest(String username, String password) {
-        System.out.println(username + password);
-        this.setUsername(username);
-        this.setPassword(password);
-    }
-
-    public String getUsername() {
-        return this.username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getPassword() {
-        return this.password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-}
