--- src/main/java/com/musictrainer/background/musictrainer/jwt/resource/JwtTokenResponse.java
@@ -1,18 +1,0 @@
-package com.musictrainer.background.musictrainer.jwt.resource;
-
-import java.io.Serializable;
-
-public class JwtTokenResponse implements Serializable {
-
-  private static final long serialVersionUID = 8317676219297719109L;
-
-  private final String token;
-
-    public JwtTokenResponse(String token) {
-        this.token = token;
-    }
-
-    public String getToken() {
-        return this.token;
-    }
-}
