--- src/main/java/com/musictrainer/background/musictrainer/users/service/UsersJpaRepository.java
@@ -1,12 +1,0 @@
-package com.musictrainer.background.musictrainer.users.service;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-
-@Repository
-public interface UsersJpaRepository extends JpaRepository<UsersData, Integer> {
-
-    List<UsersData> findByName(String name);
-}
