--- src/main/java/com/obsolete/security/basicauth/AuthenticationBean.java
@@ -1,25 +1,0 @@
-package com.obsolete.security.basicauth;
-
-public class AuthenticationBean {
-
-    private String message;
-
-    public String getMessage() {
-        return message;
-    }
-
-    public void setMessage(String message) {
-        this.message = message;
-    }
-
-    @Override
-    public String toString() {
-        return "HelloWorldBean{" +
-                "message='" + message + '\'' +
-                '}';
-    }
-
-    public AuthenticationBean(String message){
-        this.message = message;
-    }
-}
