--- src/main/java/com/obsolete/security/basicauth/BasicAuthenticationController.java
@@ -1,16 +1,0 @@
-package com.obsolete.security.basicauth;
-
-import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@CrossOrigin(origins = "http://localhost:4500")
-public class BasicAuthenticationController {
-
-    @GetMapping("/basicauth")
-    public AuthenticationBean helloWorldBean() {
-        return new AuthenticationBean("You are authenticated");
-    }
-}
-
