+++ src/main/resources/application.properties
@@ -2,8 +2,8 @@ logging.level.org.springframework=debug
 spring.http.log-request-details=true
 
 #
-#spring.security.user.name=David
-#spring.security.user.password=test
+spring.security.user.name=David
+spring.security.user.password=test
 
 
 ##h2 datbase url
